Renamed solution and minor changes
