+++ samples/csharp/getting-started/Ranking_Web/WebRanking/DataStructures/SearchResultData.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.ML.Data;
 
-namespace PersonalizedRanking.DataStructures
+namespace WebRanking.DataStructures
 {
     public class SearchResultData
     {
