+++ samples/csharp/getting-started/Ranking_Web/WebRanking/DataStructures/SearchResultPrediction.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace PersonalizedRanking.DataStructures
+namespace WebRanking.DataStructures
 {
     // Representation of the prediction made by the model (e.g. ranker).
     public class SearchResultPrediction
@@ -8,7 +8,7 @@ namespace PersonalizedRanking.DataStructures
 
         public uint Label { get; set; }
 
-        // Prediction made by the model that is used to indicate the relative ranking of the benchmark data instances.
+        // Prediction made by the model that is used to indicate the relative ranking of the candidate search results.
         public float Score { get; set; }
 
         // Values that are influential in determining the relevance of a data instance. This is a vector that contains concatenated columns from the underlying dataset.
