+++ samples/csharp/v1.0.0-All-Samples.sln
@@ -129,9 +129,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TFImageClassififcationE2E.S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TensorFlowImageClassification", "end-to-end-apps\DeepLearning_ImageClassification_TensorFlow\TensorFlowImageClassification\TensorFlowImageClassification.csproj", "{C5D5BEBF-DC10-4065-A27B-AB56E1ABCA47}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PersonalizedRanking.Solution", "PersonalizedRanking.Solution", "{B76DD928-A78E-497C-BA7D-83C5864452F9}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WebRanking.Solution", "WebRanking.Solution", "{B76DD928-A78E-497C-BA7D-83C5864452F9}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PersonalizedRanking", "getting-started\Ranking_PersonalizedSort\PersonalizedRanking\PersonalizedRanking.csproj", "{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebRanking", "getting-started\Ranking_Web\WebRanking\WebRanking.csproj", "{4EA790BB-76C7-471A-ADE4-6FBD183C461B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -429,14 +429,14 @@ Global
 		{C5D5BEBF-DC10-4065-A27B-AB56E1ABCA47}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C5D5BEBF-DC10-4065-A27B-AB56E1ABCA47}.Release|x64.ActiveCfg = Release|Any CPU
 		{C5D5BEBF-DC10-4065-A27B-AB56E1ABCA47}.Release|x64.Build.0 = Release|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Debug|x64.Build.0 = Debug|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Release|x64.ActiveCfg = Release|Any CPU
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A}.Release|x64.Build.0 = Release|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Debug|x64.Build.0 = Debug|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Release|x64.ActiveCfg = Release|Any CPU
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -478,7 +478,7 @@ Global
 		{EA9E37C6-8C62-4370-A9CF-369D002B89B6} = {7C3A7DA5-CBEB-420F-B7AC-CDE34BE2D52E}
 		{F2C0FCE9-9F76-4318-826E-892441E4A169} = {EF9F8695-25DE-4FE4-894A-6DE24E0BDD73}
 		{C5D5BEBF-DC10-4065-A27B-AB56E1ABCA47} = {F59681C2-D829-4538-A41A-568F7A7D07FD}
-		{97ACFE55-AC1F-41D0-B1A7-A0D97440346A} = {B76DD928-A78E-497C-BA7D-83C5864452F9}
+		{4EA790BB-76C7-471A-ADE4-6FBD183C461B} = {B76DD928-A78E-497C-BA7D-83C5864452F9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {98369941-33DD-450C-A410-B9A91C8CDE91}
